backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
APScheduler==3.11.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.0.3
huggingface-hub==0.26.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
packaging==24.2
pillow==11.0.0
playwright==1.47.0
propcache==0.2.0
psutil==6.0.0
pycparser==2.22
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
soundfile==0.12.1
SQLAlchemy==2.0.36
sympy==1.13.1
toml==0.10.2
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.17.2
